Add Tailwind to Test Site and e2e-tests (#422)
This PR configures Tailwind in our test site and end-to-end tests. We
also re-introduce screenshots for Darwin that we turned off in (#412).
Note that the `loadingOverlay` wait was fixed in our
`StudioPlaywrightPage`. Before, we waited for a button to be visible as
a check for the overlay to be gone. However, Playwright only considers
`hidden` [elements to be not
visible](https://playwright.dev/docs/actionability#visible), while our
`loadingOverlay` uses `opacity` to "toggle" visibility. Additionally,
since Playwright [defaults to fast-forwarding
transitions](https://playwright.dev/docs/api/class-pageassertions) for
`toHaveScreenshot`, we shouldn't have to worry about the opacity
transition time.

I also squeeze in a quick variable rename from `userStyles` to
`activeStyle` to more accurately describe that we are not injecting all
user styles, but only the ones on the active page and its component
tree.

J=SLAP-2912
TEST=auto, manual