Default to Running Studio Without Strict Mode (#426)
This PR adds a flag that will render Studio in React Strict Mode. Now,
Studio runs without React Strict Mode by default, but can be specified
to run in Strict Mode by adding `strict` as an argument to the empty
`--` flag (i.e. `npx studio -- strict`). This was also added to the
`dev` command in the test site, but not the e2e-tests.

Since we are trying to access an env variable to determine a conditional
render in Studio client-side, we cannot access `process.env` directly.
Vite offers `import.meta.env` as a way for client-side modules to access
env vars, but they must be prefixed with `VITE_`. This is why the env
var is called `VITE_STUDIO_STRICT`.

J=SLAP-2966
TEST=manual